refactor(#27) Made partition intersection cleaner.
The partition intersection code was cleaned to make it only
compute over two subpartitions, instead of computing the overlap
of a single partition with all subpartitions in another partition.

Thus, the logic over the entire other partition is moved to calling
code and we needed to add an iterator to help loop over cartesian
coordinates.