refactor(#27) Updated Transpose module to use mpi partition.
Updated the transpose modeule to use the MPIPartition defined
in #26.  Now, the function depends on MPI but the module depends
only on interfaces defined in the backend of the partition.

Also, the transpose now works if the input and output partitions
have different dimensions and different total numbers of ranks.